refactor flutter/cpp/fluter/main.h
