+++ flutter/cpp/flutter/dart_run_benchmark.cc
@@ -1,4 +1,4 @@
-#include "./main.h"
+#include "./dart_run_benchmark.h"
 
 #include <google/protobuf/text_format.h>
 
@@ -7,7 +7,6 @@
 #include <mutex>
 
 #include "flutter/cpp/backends/external.h"
-#include "flutter/cpp/c/backend_c.h"
 #include "flutter/cpp/datasets/ade20k.h"
 #include "flutter/cpp/datasets/coco.h"
 #include "flutter/cpp/datasets/imagenet.h"
@@ -15,38 +14,6 @@
 #include "flutter/cpp/mlperf_driver.h"
 #include "flutter/cpp/proto/backend_setting.pb.h"
 #include "flutter/cpp/proto/mlperf_task.pb.h"
-
-// On iOS we link backend statically and don't export any functions
-// Linker sees that we don't use some of the functions and removes them
-// (when building for release of profile modes)
-// But we still want to access these functions dynamically,
-// so we need fake_calls() function to prevent linker from removing them.
-extern "C" void fake_calls() {
-  volatile intptr_t a = 1;
-  if (a) return;
-  a = (intptr_t)dart_ffi_run_benchmark;
-  a = (intptr_t)dart_ffi_run_benchmark_free;
-  a = (intptr_t)dart_ffi_mlperf_config;
-  a = (intptr_t)dart_ffi_mlperf_config_free;
-  a = (intptr_t)dart_ffi_backend_match;
-  a = (intptr_t)dart_ffi_backend_match_free;
-#ifdef __APPLE__
-  a = (intptr_t)mlperf_backend_matches_hardware;
-  a = (intptr_t)mlperf_backend_create;
-  a = (intptr_t)mlperf_backend_vendor_name;
-  a = (intptr_t)mlperf_backend_accelerator_name;
-  a = (intptr_t)mlperf_backend_name;
-  a = (intptr_t)mlperf_backend_delete;
-  a = (intptr_t)mlperf_backend_issue_query;
-  a = (intptr_t)mlperf_backend_flush_queries;
-  a = (intptr_t)mlperf_backend_get_input_count;
-  a = (intptr_t)mlperf_backend_get_input_type;
-  a = (intptr_t)mlperf_backend_set_input;
-  a = (intptr_t)mlperf_backend_get_output_count;
-  a = (intptr_t)mlperf_backend_get_output_type;
-  a = (intptr_t)mlperf_backend_get_output;
-#endif
-}
 
 static ::mlperf::mobile::MlperfDriver* global_driver = nullptr;
 static ::std::mutex global_driver_mutex;
