+++ flutter/cpp/flutter/ios_linker_workaround.h
@@ -1,0 +1,9 @@
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void ios_linker_workaround();
+
+#ifdef __cplusplus
+}
+#endif
