+++ flutter/ios/Runner/Runner-Bridging-Header.h
@@ -1,2 +1,2 @@
-#include "../../cpp/flutter/main.h"
+#include "../../cpp/flutter/ios_linker_workaround.h"
 #import "GeneratedPluginRegistrant.h"
